@@ -1,10 +1,14 @@
 __version__ = "0.1.5dev0"
 
-from . import model
-from . import constants
-from . import graph
-from . import ontology
 from . import algorithm
 from . import annotations
-from . import validate
+from . import constants
+from . import graph
+from . import model
+from . import ontology
 from . import util
+from . import validate
+
+from .graph import OntologyGraph, GraphAware
+from .model import TermId, Term, MinimalTerm, Synonym, SynonymType, SynonymCategory
+from .ontology import Ontology, MinimalOntology
