@@ -1,3 +1,4 @@
 from ._api import OntologyGraph, GraphAware, OWL_THING
-from ._csr_graph import SimpleCsrOntologyGraph
-from ._factory import GraphFactory, CsrGraphFactory
+from ._csr_graph import SimpleCsrOntologyGraph  # REMOVE(v1.0.0)
+from ._factory import CsrGraphFactory  # REMOVE(v1.0.0)
+from ._factory import GraphFactory, IncrementalCsrGraphFactory
