@@ -1,1 +1,2 @@
-from ._csr import CsrMatrixBuilder, ImmutableCsrMatrix
+from ._csr import CsrMatrixBuilder  # REMOVE(v1.0.0)
+from ._csr import ImmutableCsrMatrix
